+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/ScrubbingModeParameters.java
@@ -15,12 +15,16 @@
  */
 package androidx.media3.exoplayer;
 
+import static androidx.media3.common.util.Assertions.checkArgument;
+
+import androidx.annotation.FloatRange;
 import androidx.annotation.Nullable;
 import androidx.media3.common.C;
 import androidx.media3.common.C.TrackType;
 import androidx.media3.common.util.UnstableApi;
 import com.google.common.collect.ImmutableSet;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
+import java.util.Objects;
 import java.util.Set;
 
 /**
@@ -42,6 +46,8 @@ public final class ScrubbingModeParameters {
    */
   public static final class Builder {
     private ImmutableSet<@TrackType Integer> disabledTrackTypes;
+    @Nullable private Double fractionalSeekToleranceBefore;
+    @Nullable private Double fractionalSeekToleranceAfter;
 
     /** Creates an instance. */
     public Builder() {
@@ -50,6 +56,8 @@ public final class ScrubbingModeParameters {
 
     private Builder(ScrubbingModeParameters scrubbingModeParameters) {
       this.disabledTrackTypes = scrubbingModeParameters.disabledTrackTypes;
+      this.fractionalSeekToleranceBefore = scrubbingModeParameters.fractionalSeekToleranceBefore;
+      this.fractionalSeekToleranceAfter = scrubbingModeParameters.fractionalSeekToleranceAfter;
     }
 
     /**
@@ -69,6 +77,37 @@ public final class ScrubbingModeParameters {
       return this;
     }
 
+    /**
+     * Sets the fraction of the media duration to use for {@link SeekParameters#toleranceBeforeUs}
+     * and {@link SeekParameters#toleranceAfterUs} when scrubbing.
+     *
+     * <p>Pass {@code null} for both values to use the {@linkplain ExoPlayer#getSeekParameters()
+     * player-level seek parameters} when scrubbing.
+     *
+     * <p>Defaults to {code null} for both values, so all seeks are exact (this may change in a
+     * future release).
+     *
+     * <p>See {@link ScrubbingModeParameters#fractionalSeekToleranceBefore} and {@link
+     * ScrubbingModeParameters#fractionalSeekToleranceAfter}.
+     *
+     * @param toleranceBefore The fraction of the media duration to use for {@link
+     *     SeekParameters#toleranceBeforeUs}, or null to use the player-level seek parameters.
+     * @param toleranceAfter The fraction of the media duration to use for {@link
+     *     SeekParameters#toleranceAfterUs}, or null to use the player-level seek parameters.
+     * @return This builder for convenience.
+     */
+    @CanIgnoreReturnValue
+    public Builder setFractionalSeekTolerance(
+        @Nullable @FloatRange(from = 0, to = 1) Double toleranceBefore,
+        @Nullable @FloatRange(from = 0, to = 1) Double toleranceAfter) {
+      checkArgument((toleranceBefore == null) == (toleranceAfter == null));
+      checkArgument(toleranceBefore == null || (toleranceBefore >= 0 && toleranceBefore <= 1));
+      checkArgument(toleranceAfter == null || (toleranceAfter >= 0 && toleranceAfter <= 1));
+      this.fractionalSeekToleranceBefore = toleranceBefore;
+      this.fractionalSeekToleranceAfter = toleranceAfter;
+      return this;
+    }
+
     /** Returns the built {@link ScrubbingModeParameters}. */
     public ScrubbingModeParameters build() {
       return new ScrubbingModeParameters(this);
@@ -78,8 +117,32 @@ public final class ScrubbingModeParameters {
   /** Which track types will be disabled in scrubbing mode. */
   public final ImmutableSet<@TrackType Integer> disabledTrackTypes;
 
+  /**
+   * The fraction of the media duration to use for {@link SeekParameters#toleranceBeforeUs} when
+   * scrubbing.
+   *
+   * <p>If this is {@code null} or the media duration is not known then the {@linkplain
+   * ExoPlayer#getSeekParameters()} non-scrubbing seek parameters} are used.
+   */
+  @Nullable
+  @FloatRange(from = 0, to = 1)
+  public final Double fractionalSeekToleranceBefore;
+
+  /**
+   * The fraction of the media duration to use for {@link SeekParameters#toleranceAfterUs} when
+   * scrubbing.
+   *
+   * <p>If this is {@code null} or the media duration is not known then the {@linkplain
+   * ExoPlayer#getSeekParameters()} non-scrubbing seek parameters} are used.
+   */
+  @Nullable
+  @FloatRange(from = 0, to = 1)
+  public final Double fractionalSeekToleranceAfter;
+
   private ScrubbingModeParameters(Builder builder) {
     this.disabledTrackTypes = builder.disabledTrackTypes;
+    this.fractionalSeekToleranceBefore = builder.fractionalSeekToleranceBefore;
+    this.fractionalSeekToleranceAfter = builder.fractionalSeekToleranceAfter;
   }
 
   /** Returns a {@link Builder} initialized with the values from this instance. */
@@ -93,11 +156,14 @@ public final class ScrubbingModeParameters {
       return false;
     }
     ScrubbingModeParameters that = (ScrubbingModeParameters) o;
-    return disabledTrackTypes.equals(that.disabledTrackTypes);
+    return disabledTrackTypes.equals(that.disabledTrackTypes)
+        && Objects.equals(fractionalSeekToleranceBefore, that.fractionalSeekToleranceBefore)
+        && Objects.equals(fractionalSeekToleranceAfter, that.fractionalSeekToleranceAfter);
   }
 
   @Override
   public int hashCode() {
-    return disabledTrackTypes.hashCode();
+    return Objects.hash(
+        disabledTrackTypes, fractionalSeekToleranceBefore, fractionalSeekToleranceAfter);
   }
 }